+++ src/componets/view.js
